+++ sys/net/network_layer/sixlowpan/sixlowpan_print.c
@@ -41,7 +41,7 @@ void sixlowpan_print(uint8_t *data, size_t size)
 
         puts("Fragmentation Header (first)");
         printf("datagram size: %" PRIu16 "\n",
-               byteorder_ntohs(hdr->disp_size) & SIXLOWPAN_FRAG_SIZE_MASK);
+               (uint16_t) (byteorder_ntohs(hdr->disp_size) & SIXLOWPAN_FRAG_SIZE_MASK));
         printf("tag: 0x%" PRIu16 "\n", byteorder_ntohs(hdr->tag));
 
         /* Print next dispatch */
@@ -53,7 +53,7 @@ void sixlowpan_print(uint8_t *data, size_t size)
 
         puts("Fragmentation Header (subsequent)");
         printf("datagram size: %" PRIu16 "\n",
-               byteorder_ntohs(hdr->disp_size) & SIXLOWPAN_FRAG_SIZE_MASK);
+               (uint16_t) (byteorder_ntohs(hdr->disp_size) & SIXLOWPAN_FRAG_SIZE_MASK));
         printf("tag: 0x%" PRIu16 "\n", byteorder_ntohs(hdr->tag));
         printf("offset: 0x%" PRIu8 "\n", hdr->offset);
 
@@ -240,7 +240,7 @@ void sixlowpan_print(uint8_t *data, size_t size)
         if (data[1] & SIXLOWPAN_IPHC2_CID_EXT) {
             offset += SIXLOWPAN_IPHC_CID_EXT_LEN;
             printf("SCI: 0x%" PRIx8 "; DCI: 0x%" PRIx8 "\n",
-                   data[2] >> 4, data[2] & 0xf);
+                   (uint8_t) (data[2] >> 4), (uint8_t) (data[2] & 0xf));
         }
 
         od_hex_dump(data + offset, size - offset, OD_WIDTH_DEFAULT);
